+++ lizard_portal/media/lizard_portal/lizard/ContextManager.js
@@ -237,18 +237,20 @@
       return output;
     },
     saveContext: function() {
-      var context, portalWindow,
+      var context, me, portalWindow,
         _this = this;
-      if (this.context && this.context.user && this.context.user.id) {
+      me = Lizard.CM;
+      if (me.context && me.context.user && me.context.user.id) {
+        console.log('Saving context...');
         context = Ext.JSON.encode({
-          objects: this.objects,
+          objects: me.objects,
           context: {
             period: {
-              start: Ext.Date.format(this.context.period.start, 'Y-m-d'),
-              end: Ext.Date.format(this.context.period.end, 'Y-m-d'),
-              type: this.context.period.type
+              start: Ext.Date.format(me.context.period.start, 'Y-m-d'),
+              end: Ext.Date.format(me.context.period.end, 'Y-m-d'),
+              type: me.context.period.type
             },
-            background_layer: this.context.background_layer
+            background_layer: me.context.background_layer
           }
         });
         portalWindow = Ext.getCmp('portalWindow');
@@ -265,7 +267,7 @@
           },
           failure: function(error) {
             console.log(error);
-            Ext.Msg.alert("Fout", "Fout in ophalen van scherm. Error: " + error);
+            Ext.Msg.alert("Fout", "Fout in het opslaan van de context. Error: " + error);
             return portalWindow.setLoading(false);
           }
         });